\name{RChoiceHB}
\alias{RChoiceHB-package}
\docType{package}
\title{
\packageTitle{RChoiceHB}
}
\description{

}

\usage{
  RChoiceHB::RChoiceHB(nRespondents, nPar, featureLengths, X_cat, Y_indexed, yIndices_old, numberOfScreens, packagesPerScreen)
}
\arguments{
  \item{nRespondents}{number of respondents}
  \item{nPar}{number of dummy variables}
  \item{featureLengths}{array of lengths of each feature}
  \item{X_cat}{X matrix}
  \item{Y_indexed}{Y matrix indexed}
  \item{yIndices_old}{Y indices}
  \item{numberOfScreens}{number of screens per respondent}
  \item{packagesPerScreen}{number of packages shown per screen}
}
\author{
\packageAuthor{RChoiceHB}
}
\keyword{ package }

